gitmodule imports and POSSE integration
add: import contentBlog as gitmodule
feat: add a ./.well-known file for POSSE integration.
POSSE stands for Post (on) Own Site Syndicate Elsewhere;
this should enable for the github.io subdomain to be used
as a handle for bluesky, as well as the main place to synchronize
every content published at the fediverse on mastodon.